Catch all exceptions when adding challenger to arena; fix legacy setData hook

+++ src/main/java/net/caseif/flint/steel/util/helper/LegacyHelper.java
@@ -46,7 +46,7 @@ public class LegacyHelper {
 
         try {
             //noinspection JavaReflectionMemberAccess
-            method_Block_setData = Block.class.getMethod("setData", short.class);
+            method_Block_setData = Block.class.getMethod("setData", byte.class);
         } catch (NoSuchMethodException e) {
             working = false;
